fix bug

+++ sys/net/gnrc/link_layer/gomach/gomach.c
@@ -1440,7 +1440,7 @@ static void gomach_send_RI_beacon(gnrc_netdev_t *gnrc_netdev)
 static void gomach_listen_init(gnrc_netdev_t *gnrc_netdev)
 {
     //puts("C");
-	printf("C:%d\n",RTT_TICKS_TO_MIN(rtt_get_counter()));
+	printf("C:%d\n",(int)RTT_TICKS_TO_MIN(rtt_get_counter()));
 
     /* Reset last_seq_info, for avoiding receiving duplicate packets.
      * To-do: remove this in the future? */
